requirements.txt
cryptography>=42.0.0
pynacl>=1.5.0
PyYAML>=6.0.1
httpx>=0.27.0
tenacity>=8.2.3
PyGithub>=2.3.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
